Create PWM module
